Conflicts:
	modules/monitoring/application/views/scripts/hosts/show.phtml

+++ modules/monitoring/application/controllers/ListController.php
@@ -1,6 +1,4 @@
 <?php
-// {{{ICINGA_LICENSE_HEADER}}}
-// {{{ICINGA_LICENSE_HEADER}}}
 
 use Icinga\Module\Monitoring\Controller;
 use Icinga\Module\Monitoring\Backend;
@@ -12,13 +10,9 @@ use Icinga\Web\Widget\Tabextension\DashboardAction;
 use Icinga\Web\Widget\Tabextension\OutputFormat;
 use Icinga\Web\Widget\Tabs;
 use Icinga\Web\Widget\SortBox;
-use Icinga\Web\Widget\FilterBox;
-use Icinga\Web\Widget\Chart\HistoryColorGrid;
 use Icinga\Data\Filter\Filter;
 use Icinga\Web\Widget;
-use Icinga\Module\Monitoring\Web\Widget\SelectBox;
 use Icinga\Module\Monitoring\Forms\StatehistoryForm;
-use Icinga\Module\Monitoring\Forms\EventOverviewForm;
 
 class Monitoring_ListController extends Controller
 {
@@ -97,12 +91,10 @@ class Monitoring_ListController extends Controller
         }
 
         // Handle soft and hard states
-        $stateType = $this->params->shift('stateType', 'soft');
-        if ($stateType == 'hard') {
+        if (strtolower($this->params->shift('stateType', 'soft')) === 'hard') {
             $stateColumn = 'host_hard_state';
             $stateChangeColumn = 'host_last_hard_state_change';
         } else {
-            $stateType = 'soft';
             $stateColumn = 'host_state';
             $stateChangeColumn = 'host_last_state_change';
         }
@@ -112,6 +104,7 @@ class Monitoring_ListController extends Controller
         $query = $this->backend->select()->from('hostStatus', array_merge(array(
             'host_icon_image',
             'host_name',
+            'host_display_name',
             'host_state' => $stateColumn,
             'host_address',
             'host_acknowledged',
@@ -138,13 +131,14 @@ class Monitoring_ListController extends Controller
 
         $this->filterQuery($query);
 
+        $this->applyRestriction('monitoring/hosts/filter', $query);
+
         $this->setupSortControl(array(
-            'host_last_check'   => $this->translate('Last Check'),
             'host_severity'     => $this->translate('Severity'),
-            'host_name'         => $this->translate('Hostname'),
-            'host_address'      => $this->translate('Address'),
             'host_state'        => $this->translate('Current State'),
-            'host_state'        => $this->translate('Hard State')
+            'host_display_name' => $this->translate('Hostname'),
+            'host_address'      => $this->translate('Address'),
+            'host_last_check'   => $this->translate('Last Check')
         ));
         $this->view->hosts = $query->paginate();
 
@@ -171,14 +165,12 @@ class Monitoring_ListController extends Controller
         }
 
         // Handle soft and hard states
-        $stateType = $this->params->shift('stateType', 'soft');
-        if ($stateType == 'hard') {
+        if (strtolower($this->params->shift('stateType', 'soft')) === 'hard') {
             $stateColumn = 'service_hard_state';
             $stateChangeColumn = 'service_last_hard_state_change';
         } else {
             $stateColumn = 'service_state';
             $stateChangeColumn = 'service_last_state_change';
-            $stateType = 'soft';
         }
 
         $this->addTitleTab('services', $this->translate('Services'));
@@ -192,6 +184,7 @@ class Monitoring_ListController extends Controller
 
         $columns = array_merge(array(
             'host_name',
+            'host_display_name',
             'host_state',
             'host_state_type',
             'host_last_state_change',
@@ -227,15 +220,17 @@ class Monitoring_ListController extends Controller
         $query = $this->backend->select()->from('serviceStatus', $columns);
 
         $this->filterQuery($query);
+
+        $this->applyRestriction('monitoring/services/filter', $query);
+
         $this->setupSortControl(array(
-            'service_last_check'    => $this->translate('Last Service Check'),
-            'service_severity'      => $this->translate('Severity'),
+            'service_severity'      => $this->translate('Service Severity'),
             'service_state'         => $this->translate('Current Service State'),
-            'service_description'   => $this->translate('Service Name'),
-            'service_state_type'    => $this->translate('Hard State'),
+            'service_display_name'  => $this->translate('Service Name'),
+            'service_last_check'    => $this->translate('Last Service Check'),
             'host_severity'         => $this->translate('Host Severity'),
             'host_state'            => $this->translate('Current Host State'),
-            'host_name'             => $this->translate('Host Name'),
+            'host_display_name'     => $this->translate('Hostname'),
             'host_address'          => $this->translate('Host Address'),
             'host_last_check'       => $this->translate('Last Host Check')
         ));
@@ -295,26 +290,31 @@ class Monitoring_ListController extends Controller
             'host'            => 'downtime_host',
             'service'         => 'downtime_service',
             'host_state'      => 'downtime_host_state',
-            'service_state'   => 'downtime_service_state'
-        ))->order('downtime_is_in_effect', 'DESC')
-          ->order('downtime_scheduled_start', 'DESC');
+            'service_state'   => 'downtime_service_state',
+            'host_display_name',
+            'service_display_name'
+        ));
 
         $this->filterQuery($query);
 
         $this->setupSortControl(array(
-            'downtime_is_in_effect'    => $this->translate('Is In Effect'),
-            'downtime_host'            => $this->translate('Host / Service'),
-            'downtime_entry_time'      => $this->translate('Entry Time'),
-            'downtime_author'          => $this->translate('Author'),
-            'downtime_start'           => $this->translate('Start Time'),
-            'downtime_start'           => $this->translate('End Time'),
-            'downtime_scheduled_start' => $this->translate('Scheduled Start'),
-            'downtime_scheduled_end'   => $this->translate('Scheduled End'),
-            'downtime_duration'        => $this->translate('Duration'),
+            'downtime_is_in_effect'     => $this->translate('Is In Effect'),
+            'host_display_name'         => $this->translate('Host'),
+            'service_display_name'      => $this->translate('Service'),
+            'downtime_entry_time'       => $this->translate('Entry Time'),
+            'downtime_author'           => $this->translate('Author'),
+            'downtime_start'            => $this->translate('Start Time'),
+            'downtime_end'              => $this->translate('End Time'),
+            'downtime_scheduled_start'  => $this->translate('Scheduled Start'),
+            'downtime_scheduled_end'    => $this->translate('Scheduled End'),
+            'downtime_duration'         => $this->translate('Duration')
         ));
 
         $this->view->downtimes = $query->paginate();
-        $this->view->delDowntimeForm = new DeleteDowntimeCommandForm();
+
+        if ($this->Auth()->hasPermission('monitoring/command/downtime/delete')) {
+            $this->view->delDowntimeForm = new DeleteDowntimeCommandForm();
+        }
     }
 
     /**
@@ -333,7 +333,9 @@ class Monitoring_ListController extends Controller
             'notification_output',
             'notification_contact',
             'notification_start_time',
-            'notification_state'
+            'notification_state',
+            'host_display_name',
+            'service_display_name'
         ));
         $this->filterQuery($query);
         $this->view->notifications = $query->paginate();
@@ -471,20 +473,26 @@ class Monitoring_ListController extends Controller
             'persistent' => 'comment_is_persistent',
             'expiration' => 'comment_expiration',
             'host'       => 'comment_host',
-            'service'    => 'comment_service'
+            'service'    => 'comment_service',
+            'host_display_name',
+            'service_display_name'
         ));
         $this->filterQuery($query);
         $this->view->comments = $query->paginate();
 
         $this->setupSortControl(
             array(
-                'comment_timestamp'  => $this->translate('Comment Timestamp'),
-                'comment_host'       => $this->translate('Host / Service'),
-                'comment_type'       => $this->translate('Comment Type'),
-                'comment_expiration' => $this->translate('Expiration'),
+                'comment_timestamp'     => $this->translate('Comment Timestamp'),
+                'host_display_name'     => $this->translate('Host'),
+                'service_display_name'  => $this->translate('Service'),
+                'comment_type'          => $this->translate('Comment Type'),
+                'comment_expiration'    => $this->translate('Expiration')
             )
         );
-        $this->view->delCommentForm = new DeleteCommentCommandForm();
+
+        if ($this->Auth()->hasPermission('monitoring/command/comment/delete')) {
+            $this->view->delCommentForm = new DeleteCommentCommandForm();
+        }
     }
 
     public function servicegroupsAction()
@@ -496,6 +504,7 @@ class Monitoring_ListController extends Controller
         $this->setAutorefreshInterval(12);
         $query = $this->backend->select()->from('groupsummary', array(
             'servicegroup',
+            'servicegroup_alias',
             'hosts_up',
             'hosts_unreachable_handled',
             'hosts_unreachable_unhandled',
@@ -519,18 +528,20 @@ class Monitoring_ListController extends Controller
             'services_critical_last_state_change_unhandled',
             'services_unknown_last_state_change_unhandled',
             'services_total'
-        ));
+        ))->order('services_severity')->order('servicegroup_alias');
+        // TODO(el): Can't default to the sort rules of the data view because it's meant for both host groups and
+        // service groups. We should separate them.
         $this->filterQuery($query);
         $this->view->servicegroups = $query->paginate();
         $this->setupSortControl(array(
-            'services_severity' => $this->translate('Severity'),
-            'servicegroup'      => $this->translate('Service Group Name'),
-            'services_total'    => $this->translate('Total Services'),
-            'services_ok'       => $this->translate('Services OK'),
-            'services_unknown'  => $this->translate('Services UNKNOWN'),
-            'services_critical' => $this->translate('Services CRITICAL'),
-            'services_warning'  => $this->translate('Services WARNING'),
-            'services_pending'  => $this->translate('Services PENDING')
+            'services_severity'     => $this->translate('Severity'),
+            'servicegroup_alias'    => $this->translate('Service Group Name'),
+            'services_total'        => $this->translate('Total Services'),
+            'services_ok'           => $this->translate('Services OK'),
+            'services_unknown'      => $this->translate('Services UNKNOWN'),
+            'services_critical'     => $this->translate('Services CRITICAL'),
+            'services_warning'      => $this->translate('Services WARNING'),
+            'services_pending'      => $this->translate('Services PENDING')
         ));
     }
 
@@ -543,6 +554,7 @@ class Monitoring_ListController extends Controller
         $this->setAutorefreshInterval(12);
         $query = $this->backend->select()->from('groupsummary', array(
             'hostgroup',
+            'hostgroup_alias',
             'hosts_up',
             'hosts_unreachable_handled',
             'hosts_unreachable_unhandled',
@@ -566,12 +578,14 @@ class Monitoring_ListController extends Controller
             'services_critical_last_state_change_unhandled',
             'services_unknown_last_state_change_unhandled',
             'services_total'
-        ));
+        ))->order('services_severity')->order('hostgroup_alias');
+        // TODO(el): Can't default to the sort rules of the data view because it's meant for both host groups and
+        // service groups. We should separate them.
         $this->filterQuery($query);
         $this->view->hostgroups = $query->paginate();
         $this->setupSortControl(array(
             'services_severity' => $this->translate('Severity'),
-            'hostgroup'         => $this->translate('Host Group Name'),
+            'hostgroup_alias'   => $this->translate('Host Group Name'),
             'services_total'    => $this->translate('Total Services'),
             'services_ok'       => $this->translate('Services OK'),
             'services_unknown'  => $this->translate('Services UNKNOWN'),
@@ -590,7 +604,9 @@ class Monitoring_ListController extends Controller
 
         $query = $this->backend->select()->from('eventHistory', array(
             'host_name',
+            'host_display_name',
             'service_description',
+            'service_display_name',
             'object_type',
             'timestamp',
             'state',
@@ -605,7 +621,7 @@ class Monitoring_ListController extends Controller
         $this->filterQuery($query);
 
         $this->setupSortControl(array(
-            'timestamp' => 'Occurence'
+            'timestamp' => $this->translate('Occurence')
         ));
         $this->view->history = $query->paginate();
     }
@@ -650,19 +666,7 @@ class Monitoring_ListController extends Controller
         if ($sort = $this->params->get('sort')) {
             $query->order($sort, $this->params->get('dir'));
         }
-        $this->applyRestrictions($query);
         $this->handleFormatRequest($query);
-        return $query;
-    }
-
-    /**
-     * Apply current user's `monitoring/filter' restrictions on the given data view
-     */
-    protected function applyRestrictions($query)
-    {
-        foreach ($this->getRestrictions('monitoring/filter') as $restriction) {
-            // TODO: $query->applyFilter(Filter::fromQueryString());
-        }
         return $query;
     }
